Correct ping test logic in pfSense-dhclient-script.
pfSense-dhclient-script fails to return nonzero in the case where a DHCP timeout occurs and the cached gateway is not pingable. This results in a case where the cached IP is removed from the interface but dhclient is informed that the rebinding was successful. As a result, the interface remains without an IPv4 address until the DHCP lease expires, instead of the DHCP protocol being restarted after the defined retry interval.